add bot

+++ tool.sh
@@ -1,12 +1,15 @@
 if [ "$1" = "make" ]; then
+	if [ ! -d "build" ]; then
+		mkdir build
+	fi
 	cd build && cmake ../ && make && cd ..
-elif [ "$1" = "clean" ]; then
+elif [ "$1" = "clean" -a -d "/build" ] ; then
 	rm Tank && cd build && rm -rf * && cd ..
 elif [ "$1" = "merge" ]; then
 	python3 merge.py
 elif [ "$1" = "format" ]; then
 	clang-format -style=file -i `find . -type f -regex ".*\.\(cpp\|h\)"`
 else
-	echo "No such command '$1'"
+	echo "ERROR"
 fi
 echo "FINISHED"
